+++ src/cli.rs
@@ -16,6 +16,7 @@ Cargo subcommand for running cargo without dev-dependencies.
 ";
 
 pub(crate) struct Args {
+    pub(crate) no_private: bool,
     pub(crate) manifest_path: Option<Utf8PathBuf>,
     pub(crate) cargo_args: Vec<String>,
     pub(crate) rest: Vec<String>,
@@ -60,6 +61,8 @@ impl Args {
         let mut manifest_path: Option<Utf8PathBuf> = None;
         let mut verbose = 0;
 
+        let mut no_private = false;
+
         let mut parser = lexopt::Parser::from_args(args);
         while let Some(arg) = parser.next()? {
             macro_rules! parse_opt {
@@ -70,11 +73,21 @@ impl Args {
                     $opt = Some(parser.value()?.parse()?);
                 }};
             }
+            macro_rules! parse_flag {
+                ($flag:ident $(,)?) => {{
+                    if $flag {
+                        multi_arg(&arg)?;
+                    }
+                    $flag = true;
+                }};
+            }
 
             match arg {
                 Long("color") => parse_opt!(color),
                 Long("manifest-path") => parse_opt!(manifest_path),
                 Short('v') | Long("verbose") => verbose += 1,
+
+                Long("no-private") => parse_flag!(no_private),
 
                 Short('h') | Long("help") if subcommand.is_none() => {
                     print!("{USAGE}");
@@ -131,7 +144,7 @@ impl Args {
             cargo_args.push(path.as_str().to_owned());
         }
 
-        Ok(Self { manifest_path, cargo_args, rest })
+        Ok(Self { no_private, manifest_path, cargo_args, rest })
     }
 }